# internal/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API for the UnderNET Channel Service. It provides a RESTful interface for managing users, channels, and other resources. \u003c!-- ReDoc-Inject: \u003csecurity-definitions\u003e --\u003e\n# Authorization\n**JWT Bearer Token:** The main authorization method for the API. Needs `Authorization: Bearer \u003cjwt-token\u003e` HTTP header to authenticate.\n\u003c!-- ReDoc-Inject: \u003csecurity-definitions\u003e --\u003e",
        "title": "UnderNET Channel Service API",
        "contact": {
            "name": "Ratler",
            "email": "ratler@undernet.org"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/UndernetIRC/cservice-api/blob/master/LICENSE"
        },
        "version": "0.1"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/admin/roles": {
            "get": {
                "security": [
                    {
                        "JWTBearerToken": []
                    }
                ],
                "description": "Returns a list of roles",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "List roles",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/admin.RoleListResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "JWTBearerToken": []
                    }
                ],
                "description": "Creates a new role",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Create role",
                "parameters": [
                    {
                        "description": "Role data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/admin.RoleDataRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/admin.RoleCreateResponse"
                        }
                    }
                }
            }
        },
        "/admin/roles/{id}": {
            "put": {
                "security": [
                    {
                        "JWTBearerToken": []
                    }
                ],
                "description": "Updates a role",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Update role",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Role ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Role data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/admin.RoleDataRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/admin.roleUpdateResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "JWTBearerToken": []
                    }
                ],
                "description": "Deletes a role",
                "tags": [
                    "admin"
                ],
                "summary": "Delete role",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Role ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/admin/roles/{id}/users": {
            "post": {
                "security": [
                    {
                        "JWTBearerToken": []
                    }
                ],
                "description": "Assigns users to a role",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Assign users to role",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Role ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "List of usernames",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/admin.UsersRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "JWTBearerToken": []
                    }
                ],
                "description": "Removes users from a role",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Remove users from role",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Role ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "List of usernames",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/admin.UsersRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "admin.RoleCreateResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "admin.RoleDataRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 3,
                    "x-order": "0"
                },
                "description": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 3,
                    "x-order": "1"
                }
            }
        },
        "admin.RoleListResponse": {
            "type": "object",
            "properties": {
                "roles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/admin.RoleNameResponse"
                    }
                }
            }
        },
        "admin.RoleNameResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "x-order": "0"
                },
                "name": {
                    "type": "string",
                    "x-order": "1"
                },
                "description": {
                    "type": "string",
                    "x-order": "2"
                }
            }
        },
        "admin.UsersRequest": {
            "type": "object",
            "required": [
                "users"
            ],
            "properties": {
                "users": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "admin.roleUpdateResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "JWTBearerToken": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}